requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-adk==0.8.0
openai==1.50.0
pymongo==4.6.1
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
mcp-client==0.7.0
jinja2==3.1.2
aiofiles==23.2.1
